+++ include/game/player.h
@@ -3,15 +3,18 @@
 
 #include "game.h"
 #include "map.h"
+#include "bullet.h"
 class Game;
 class Map;
+class Bullet;
 
 class Player {
 private:
     Game* game;
     Map* map;
 
-    LevelPos pos = { 0, 0 };
+    LevelPos pos;
+    std::vector<Bullet*> bullets;
 
 public:
     Player(Game* game, Map* map);
@@ -21,6 +24,8 @@ public:
     void render();
 
     void shot(LevelPos pos);
+
+    std::vector<Bullet*>* getBullets();
 };
 
 #endif