+++ include/game/utils.h
@@ -7,12 +7,31 @@ enum GameState {
     PAUSE,
 };
 
+enum ItemType {
+    GROUND = 0,
+    WALL = 1,
+    HOLE = 2,
+};
+
 struct LevelPos {
     float x, y;
 };
 
 struct LevelSize {
     float w, h;
+};
+
+struct LevelRect {
+    float x, y, w, h;
+
+    bool intersect(LevelPos pos) {
+        float left = x - w / 2.0f;
+        float right = x + w / 2.0f;
+        float top = y - h / 2.0f;
+        float bottom = y + h / 2.0f;
+
+        return (pos.x >= left && pos.x <= right && pos.y >= top && pos.y <= bottom);
+    }
 };
 
 struct Rotation {
@@ -37,10 +56,20 @@ struct LevelData {
     unsigned int id;
 };
 
-struct KeyEvent {
-    SDL_KeyCode keycode;
-    GameState state;
-    std::function<void()> call;
+struct BulletData {
+    float speed = 1.0f;
+    float damage = 1.0f;
+    int rebound = 0;
+    bool pierce = false;
+    bool explosive = false;
+    Transform tranform;
+};
+
+struct Item {
+    ItemType type;
+    LevelPos pos;
+
+    LevelRect getRect() { return { pos.x - 0.5f, pos.y -0.5f, 1.0f, 1.0f }; }
 };
 
 #endif