+++ src/camera.cpp
@@ -2,9 +2,8 @@
 
 // INIT
 Camera::Camera(Game* g) : game(g) {
-    // static GameData* gameData = game->getData();
-    
-    // setPos(gameData->positions->camera, 0);
+    static GameData* gameData = game->getData();
+    setPos(gameData->positions->camera);
 }
 
 // CLEANUP
@@ -17,7 +16,7 @@ LevelPos Camera::getPos() {
     return pos;
 }
 
-//SETTER
+// SETTER
 void Camera::setPos(LevelPos p) {
     pos = p;
 }