+++ CHANGES.md
@@ -1,3 +1,9 @@
+## HB-1.0.1
+#### ADDED
+- bullets shooting
+- camera movements
+- map generation
+
 ## HB-1.0.0
 #### ADDED
 - player movements